ntools_posvol

Struct Dimensions

pub struct Dimensions {
    pub res_x: i32,
    pub res_y: i32,
    pub res_z: i32,
    pub n_x: i32,
    pub n_y: i32,
    pub n_z: i32,
}

Dimension values in the first Posvol data block

Stores the six dimension values in the first block of binary data.

Fields correspond to the sample resolution in x, y, and z dimensions, and the number of mesh bounds in each mesh coordinate axis.

Fields
res_x: i32

Sample resolution in x

res_y: i32

Sample resolution in y

res_z: i32

Sample resolution in z

n_x: i32

Number of mesh bounds in x (iints+1)

n_y: i32

Number of mesh bounds in y (iints+1)

n_z: i32

Number of mesh bounds in z (iints+1)

Implementations

impl Dimensions

pub fn number_of_voxels(&self) -> usize

Number of voxels expected in the file

Taken from the mesh bounds found in the dimensions of the file header.

pub fn number_of_subvoxels(&self) -> usize

Number of samples per voxel expected in the file

For example: If the sample resolution is 3x4x5, then there should be 60 regions (sub-voxels) inside every voxel.

pub fn number_of_cells(&self) -> usize

Total number of cells expected in the file

The product of the number of voxels and the number of samples per voxel.

For example: If the sample resolution is 3x4x5 over a mesh with 100 voxels, then there will be 100*60 total cells. i.e. one per sub-voxel region.

pub fn cell_array_byte_length(&self) -> usize

Expected size of full cells array based on the header dimensions